+++ cli/index.js
@@ -13,6 +13,7 @@ import { infoCommand } from './commands/info.js';
 import { verifyCommand } from './commands/verify.js';
 import { updateCommand } from './commands/update.js';
 import { manifestCommand } from './commands/manifest.js';
+import { enrichCommand } from './commands/enrich.js';
 
 // Load package.json to get version
 const packageJson = await fs.readJson(new URL('../package.json', import.meta.url));
@@ -35,5 +36,6 @@ program.addCommand(infoCommand());
 program.addCommand(verifyCommand());
 program.addCommand(updateCommand());
 program.addCommand(manifestCommand());
+program.addCommand(enrichCommand());
 
 program.parse();